cargasB3/requirements.txt
requests>=2.0.0
azure-storage-blob>=12.0.0
azure-cosmos>=4.0.0
lxml
SQLAlchemy>=2.0
mysql-connector-python
flask
flask-cors
gunicorn
azure-functions
